cdvae/cdvae-main/requirements.txt
ase==3.22
autopep8
# jupyterlab
matminer==0.7.3
matplotlib
nglview
# ipywidgets
pylint
# pymatgen==2020.12.31
pymatgen
pytorch-lightning==1.3.8
# pytorch-lightning
seaborn
tqdm
hydra-core==1.1.0
hydra-joblib-launcher==1.1.5
p-tqdm==1.3.3
pytest
python-dotenv
smact==2.2.1
streamlit==0.79.0
wandb==0.10.33
# wandb
numpy==1.25.0
protobuf==3.19.6
torch_geometric==1.7.2
# pip install torch-scatter==2.0.7 -f https://data.pyg.org/whl/torch-1.8.0+cu102.html
# pip install torch-sparse==0.6.10 -f https://data.pyg.org/whl/torch-1.8.0+cu102.html
# pip install torch-cluster==1.5.9 -f https://data.pyg.org/whl/torch-1.8.0+cu102.html
# pip install torch-spline-conv==1.2.1 -f https://data.pyg.org/whl/torch-1.8.0+cu102.html
# pip install torch-geometric==1.7.2
